payaki events and training get order id on based on that id call payment apis

+++ lib/network/model/response/payment/payment_done_response.dart
@@ -29,12 +29,12 @@ class Payment {
   String? merchantTransactionId;
   String? type;
   String? operation;
-  int? amount;
+  double? amount;
   String? currency;
   String? status;
   String? description;
   bool? disputes;
-  int? applicationFeeAmount;
+  double? applicationFeeAmount;
   String? paymentMethod;
   DateTime? createdDate;
   DateTime? updatedDate;
